+++ src/manager/errorManager.js
@@ -6,7 +6,7 @@ const ErrorManager = {
         case 'EBUSY':
           return 'That file may be open in another application. Close it, and try again.';
         case 'EPERM':
-          return 'You are missing permissions to modify some files. Try running Minecraft Manager as the administrator.';
+          return `You are missing permissions to modify some files. Try running Minecraft Manager as the administrator. ${error.toString()}`;
         default:
           return error.toString();
       }